Cameras Fixed
Cameras now transition smoothly between areas.
Current location of the player is stored as a string & updated when the
player changes areas.
WASD movement for the player is disabled during certain area transitions
where camera-based movement can be problematic.
Added a plane so the player can't jump into the library from the
hallway.

In the next update:
Will import Regina's world assets as available
Will use a graphical trick to hide the library in blackness when on the
second floor (method tbd)
Will be pretty cool B)